+++ NEWS.md
@@ -1,3 +1,13 @@
+# wbpip (development version)
+
+* Add new functions to compute the following indicators:  
+  - Number of poor
+  - Average shortfall
+  - Total shortfall
+  - Income gap ratio
+  - Palma ratio
+  - Top 10 / Bottom 10 ratio
+
 # wbpip 0.1.0
 
 ## New features